pyproject.toml
[project]
name = "aws-textract-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amazon-textract-response-parser>=1.0.3",
    "amazon-textract-textractor>=1.9.2",
    "boto3>=1.42.82",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.14",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "pdf2image>=1.17.0",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
]
